Added Color Swatch addition and resorting behaviour
The Color Swatch accepts new colors until the swatch is full. After
this, new colors will force the oldest color to shift off the array to
make room at the end.

If a color already exists in the array it is simply pushed to the end of
the array to become the 'Most Recent'